pyproject.toml
[project]
name = "keralajersey-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "appwrite>=14.1.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
